+++ server/service/Login.php
@@ -58,6 +58,10 @@ class Login
         $this->init_session();
     }
 
+    /**
+     * Check the default user locale and if we have the same we assign it.
+     * If there is not the same we check for the same language not locale.
+     */
     private function use_user_locale(){        
         $locale = Locale::acceptFromHttp($_SERVER['HTTP_ACCEPT_LANGUAGE']);
         // Replace underscores with hyphens
@@ -112,7 +116,6 @@ class Login
         if(!isset($_SESSION['cms_gender'])) $_SESSION['cms_gender'] = MALE_GENDER_ID;
         if(!isset($_SESSION['language'])) $_SESSION['language'] = $this->db->get_default_language();
         if(!isset($_SESSION['user_language'])) $_SESSION['user_language'] = LANGUAGE;
-        
         $this->use_user_locale();
 
         if(!isset($_SESSION['cms_language'])) $_SESSION['cms_language'] = 2;